管理系统/learning-assistant/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-jose